Factor out common conversions to/from Google objects
